Backpropagation included in multilayer-perceptron
