+++ docs/index.md
@@ -65,8 +65,7 @@ Example scripts are available in a separate repository [php-ai/php-ml-examples](
 * Workflow
     * [Pipeline](machine-learning/workflow/pipeline)
 * Neural Network
-    * [Multilayer Perceptron](machine-learning/neural-network/multilayer-perceptron/)
-    * [Backpropagation training](machine-learning/neural-network/backpropagation/)
+    * [Multilayer Perceptron Classifier](machine-learning/neural-network/multilayer-perceptron-classifier/)
 * Cross Validation
     * [Random Split](machine-learning/cross-validation/random-split/)
     * [Stratified Random Split](machine-learning/cross-validation/stratified-random-split/)
