+++ src/Phpml/NeuralNetwork/Network/MultilayerPerceptron.php
@@ -4,7 +4,9 @@ declare(strict_types=1);
 
 namespace Phpml\NeuralNetwork\Network;
 
+use Phpml\Estimator;
 use Phpml\Exception\InvalidArgumentException;
+use Phpml\NeuralNetwork\Training\Backpropagation;
 use Phpml\NeuralNetwork\ActivationFunction;
 use Phpml\NeuralNetwork\Layer;
 use Phpml\NeuralNetwork\Node\Bias;
@@ -13,24 +15,36 @@ use Phpml\NeuralNetwork\Node\Neuron;
 use Phpml\NeuralNetwork\Node\Neuron\Synapse;
 use Phpml\Helper\Predictable;
 
-class MultilayerPerceptron extends LayeredNetwork
+abstract class MultilayerPerceptron extends LayeredNetwork implements Estimator
 {
     use Predictable;
 
     /**
      * @var array
      */
-    private $classes = [];
+    protected $classes = [];
+
+    /**
+     * @var int
+     */
+    private $iterations;
+
+    /**
+     * @var Backpropagation
+     */
+    protected $backpropagation = null;
 
     /**
      * @param int                     $inputLayerFeatures
      * @param array                   $hiddenLayers
      * @param array                   $classes
+     * @param int                     $iterations
      * @param ActivationFunction|null $activationFunction
+     * @param int                     $theta
      *
      * @throws InvalidArgumentException
      */
-    public function __construct(int $inputLayerFeatures, array $hiddenLayers, array $classes, ActivationFunction $activationFunction = null)
+    public function __construct(int $inputLayerFeatures, array $hiddenLayers, array $classes, int $iterations = 10000, ActivationFunction $activationFunction = null, int $theta = 1)
     {
         if (empty($hiddenLayers)) {
             throw InvalidArgumentException::invalidLayersNumber();
@@ -40,7 +54,9 @@ class MultilayerPerceptron extends LayeredNetwork
         if ($nClasses < 2) {
             throw InvalidArgumentException::invalidClassesNumber();
         }
-        $this->classes = $classes;
+        $this->classes = array_values($classes);
+
+        $this->iterations = $iterations;
 
         $this->addInputLayer($inputLayerFeatures);
         $this->addNeuronLayers($hiddenLayers, $activationFunction);
@@ -48,39 +64,32 @@ class MultilayerPerceptron extends LayeredNetwork
 
         $this->addBiasNodes();
         $this->generateSynapses();
+
+        $this->backpropagation = new Backpropagation($theta);
     }
 
     /**
-     * @param  mixed $target
-     * @return int
+     * @param array $samples
+     * @param array $targets
      */
-    public function getTargetClass($target): int
+    public function train(array $samples, array $targets)
     {
-        if (!in_array($target, $this->classes)) {
-            throw InvalidArgumentException::invalidTarget($target);
+        for ($i = 0; $i < $this->iterations; ++$i) {
+            $this->trainSamples($samples, $targets);
         }
-        return array_search($target, $this->classes);
     }
 
     /**
      * @param array $sample
-     *
+     * @param mixed $target
+     */
+    protected abstract function trainSample(array $sample, $target);
+
+    /**
+     * @param array $sample
      * @return mixed
      */
-    public function predictSample(array $sample)
-    {
-        $output = $this->setInput($sample)->getOutput();
-
-        $predictedClass = null;
-        $max = 0;
-        foreach ($output as $class => $value) {
-            if ($value > $max) {
-                $predictedClass = $class;
-                $max = $value;
-            }
-        }
-        return $this->classes[$predictedClass];
-    }
+    protected abstract function predictSample(array $sample);
 
     /**
      * @param int $nodes
@@ -140,6 +149,17 @@ class MultilayerPerceptron extends LayeredNetwork
     {
         foreach ($currentLayer->getNodes() as $currentNeuron) {
             $nextNeuron->addSynapse(new Synapse($currentNeuron));
+        }
+    }
+
+    /**
+     * @param array $samples
+     * @param array $targets
+     */
+    private function trainSamples(array $samples, array $targets)
+    {
+        foreach ($targets as $key => $target) {
+            $this->trainSample($samples[$key], $target);
         }
     }
 }
